More precise filename parsing in python module. Closes #114.
Previously, python module would check if filename started with "tcp"
or "http." This would cause weird edge cases where a relative path
starting with a directory called "tcp" or "http" would cause an
exception to be thrown. This checks for filenames that start with
"tcp://" or "http://" instead, which *might* still cause the same
issue but is less likely.