Improve error handling to avoid infinite loop
If the subprocess crashed, we got stuck in an infinite loop because of the bare except clause. This change handles SIGINT instead of ignoring it, which allows the user to press ctrl-c to escape the infinite loop. This change also checks the return code of the subprocess and raises an exception if the process segfaulted. This prevents an infinite loop where we would try to read from a crashed process forever.